# openapi.yaml
openapi: 3.1.0
info:
  title: assessment-service
  version: v0.3.1
paths:
...
  /v1/admin/assignments:
    post:
      tags:
      - assessment
      - assessment
      summary: Assign Test
      operationId: postAdminAssignments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTestRequest'
            example:
              test_id: 00000000-0000-4000-8000-000000000000
              child_id: 00000000-0000-4000-8000-000000000000
              retake: true
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
              example:
                assignment_id: 00000000-0000-4000-8000-000000000000
                test_id: 00000000-0000-4000-8000-000000000000
                child_id: 00000000-0000-4000-8000-000000000000
                status: string
                attempt_no: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/assignments:
    post:
      tags:
      - assessment
      - assessment
      summary: Assign Test
      operationId: postAssignments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTestRequest'
            example:
              test_id: 00000000-0000-4000-8000-000000000000
              child_id: 00000000-0000-4000-8000-000000000000
              retake: true
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
              example:
                assignment_id: 00000000-0000-4000-8000-000000000000
                test_id: 00000000-0000-4000-8000-000000000000
                child_id: 00000000-0000-4000-8000-000000000000
                status: string
                attempt_no: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
...
components:
  schemas:
    AssignTestRequest:
      properties:
        test_id:
          type: string
          format: uuid
          title: Test Id
        child_id:
          type: string
          format: uuid
          title: Child Id
        retake:
          type: boolean
          title: Retake
          default: false
      type: object
      required:
      - test_id
      - child_id
      title: AssignTestRequest
...
    AssignmentResponse:
      properties:
        assignment_id:
          type: string
          format: uuid
          title: Assignment Id
        test_id:
          type: string
          format: uuid
          title: Test Id
        child_id:
          type: string
          format: uuid
          title: Child Id
        status:
          type: string
          title: Status
        attempt_no:
          type: integer
          title: Attempt No
      type: object
      required:
      - assignment_id
      - test_id
      - child_id
      - status
      - attempt_no
      title: AssignmentResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ProblemDetails:
      type: object
      title: ProblemDetails
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        code:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/authentication
            title: Unauthorized
            status: 401
            detail: Bearer access token is required
            code: UNAUTHORIZED
            request_id: 00000000-0000-4000-8000-000000000000
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/access-denied
            title: Forbidden
            status: 403
            detail: Access denied
            code: FORBIDDEN
            request_id: 00000000-0000-4000-8000-000000000000
    TooManyRequests:
      description: Too Many Requests
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/too-many-requests
            title: Too Many Requests
            status: 429
            detail: Rate limit exceeded
            code: RATE_LIMITED
            request_id: 00000000-0000-4000-8000-000000000000
servers:
- url: http://localhost:8003
  description: Local
- url: https://api.monitoring.example
  description: Production
security:
- bearerAuth: []